client_genivi: add --force-display and --reuse-sessions options so that demo can work with the AGL environment
The CES 2016 AGL demo platform requires all weston surfaces (navi and qml)
to be intialized fromp the beginning. After that, we use 'LayerManagerControl'
to set them in the correct order and set them visible.
This is a problem for the 'client_genivi' process, which initializes its
weston surface only during it's lifetime (on CreateMapviewInstance()),
so too late for us.
In order to avoid this problem, the '--force-display' option allows to
pre-create the sessions (navicore, mapviewer and mapviewinstance). This
way, we can re-arrange the surfaces with 'LayerManagerControl' as usual
before starting the demo. Then, 'client_genivi' will have to be launched
with option '--reuse-sessions', so that pre-created sessions will be re-used
and not re-created. Sessions will not be deleted either.